# sort: implement numeric string comparison