--- a/tests/by-util/test_sort.rs
+++ b/tests/by-util/test_sort.rs
@@ -19,11 +19,7 @@ fn test_multiple_decimals_general() {
 
 #[test]
 fn test_multiple_decimals_numeric() {
-    new_ucmd!()
-        .arg("-n")
-        .arg("multiple_decimals_numeric.txt")
-        .succeeds()
-        .stdout_is("-2028789030\n-896689\n-8.90880\n-1\n-.05\n\n\n\n\n\n\n\n\n000\nCARAvan\n00000001\n1\n1.040000000\n1.444\n1.58590\n8.013\n45\n46.89\n               4567.\n4567.1\n4567.34\n\t\t\t\t\t\t\t\t\t\t4567..457\n\t\t\t\t37800\n\t\t\t\t\t\t45670.89079.098\n\t\t\t\t\t\t45670.89079.1\n576,446.88800000\n576,446.890\n4798908.340000000000\n4798908.45\n4798908.8909800\n");
+    test_helper("multiple_decimals_numeric", "-n")
 }
 
 #[test]