This implements -n and -h using a string comparison algorithm instead
of parsing each number to a f64 and comparing those.

This should result in a moderate performance increase and eliminate loss
of precision.

Some benchmarks that show performance improvements:

|     | before | after |comparison to gnu
|---|---|---|---
| -n  | Time (mean ± σ):      44.9 ms ±   4.4 ms    \[User: 371.9 ms, System: 12.0 ms\] | Time (mean ± σ):      25.8 ms ±   3.5 ms    \[User: 104.7 ms, System: 11.2 ms\] | 'target/release/coreutils sort shuffled_numbers.txt -n -o output.txt' ran 1.31 ± 0.06 times faster than 'LC_ALL=C sort shuffled_numbers.txt -n -o output.txt' |
| -h  | Time (mean ± σ):      67.4 ms ±   1.5 ms    \[User: 698.5 ms, System: 11.3 ms\] | Time (mean ± σ):      27.5 ms ±   1.0 ms    [User: 104.2 ms, System: 10.6 ms] | 'target/release/coreutils sort shuffled_numbers_si.txt -h -o output.txt' ran 2.52 ± 0.09 times faster than 'LC_ALL=C sort shuffled_numbers_si.txt -h -o output.txt'