@electricboogie 2021-04-11
This is very cool.  Impressed it is so concise.  One thing this made me think of is that we accept all is_numeric, instead of is_ascii_digit, at leading_num_common and perhaps we don't need to do that, conversely you could accept all is_numeric and compare all those unicode numerics. That would definitely be new behavior but perhaps its behavior that people want? Maybe an additional long opt for that case?  I know the Chinese use the Arabic numeral system, of course, but also use two indigenous systems.