@miDeb 2021-04-11
I don't think this would work for anything else than ascii digits because the individual characters are compared by their value and I think that doesn't work if we allow other digits.
(I don't know how other numeric systems work, so this might be incorrect)
I don't think they'd easily work with the assumptions that are baked into this algorithm. Those are that the nth digit of one number is as significant as the nth digit of the other one, and that the number with more digits is greater. This means that hex comparisons wouldn't work if mixed with other bases, which we'd certainly need to support.

Feel free to kick it around as much as you want ;)