@electricboogie 2021-04-11
> I don't think this would work for anything else than ascii digits because the individual characters are compared by their value and I think that doesn't work if we allow other digits.

So you're saying because they are compared by their code point converting among mixed numeral inputs would be no fun? That's fair.  I am probably just gilding the lily re: is_numeric. 

I guess what I was thinking, and I may be way off base (pun intended), is that there are a number of base 10/Arabic systems with same values but different scripts (see Middle Eastern Arabic and many of the Indian systems).  And hex to decimal conversion isn't that bad? Also I'm not sure converting between bases is required behavior.

In any event, this looks cool as is.  I'm going to try it soon.