@miDeb 2021-04-11
> Also I'm not sure converting between bases is required behavior.

It's not, only for `-g`, which explicitly asks for converting to a `f64` so I left that part unchanged.