@miDeb 2021-04-11
I added an optimization for general numeric compare (cache parsed floats). This results in a nice performance win:
Before:
Time (mean ± σ):      40.9 ms ±   1.0 ms    [User: 345.1 ms, System: 12.7 ms]
After:
Time (mean ± σ):      28.1 ms ±   1.3 ms    [User: 71.9 ms, System: 16.7 ms]